Split GeneratorParameters to another file
